:sparkles: Error messages are clear and informative | fixed lint errors
