+++ components/auth-button.vue
@@ -1,25 +1,12 @@
 <script setup lang="ts">
-import { createAuthClient } from "better-auth/client";
-
-const authClient = createAuthClient();
-const loading = ref(false);
-
-async function signIn() {
-  loading.value = true;
-  await authClient.signIn.social({
-    provider: "google",
-    callbackURL: "/dashboard",
-  });
-  loading.value = false;
-}
+const authStore = useAuthStore();
 </script>
 
 <template>
-  <button :disabled="loading" class="btn btn-accent" @click="signIn">
-    <span v-if="loading" class="text-gray-950 loading loading-dots loading-md" />
-    <template v-else>
-      Sign In With Google
-      <Icon name="tabler:brand-google-filled" size="24" />
-    </template>
+  <button :disabled="authStore.loading" class="btn btn-accent" @click="authStore.signIn">
+    Sign In With Google
+
+    <span v-if="authStore.loading" class="loading loading-spinner loading-md" />
+    <Icon v-else name="tabler:brand-google-filled" size="24" />
   </button>
 </template>
