+++ package.json
@@ -18,11 +18,13 @@
     "@nuxt/eslint": "^1.3.0",
     "@nuxt/icon": "^1.12.0",
     "@nuxtjs/color-mode": "^3.5.2",
+    "@pinia/nuxt": "^0.11.0",
     "@tailwindcss/vite": "^4.1.5",
     "better-auth": "^1.2.8",
     "concurrently": "^9.1.2",
     "drizzle-orm": "^0.43.1",
     "nuxt": "^3.16.2",
+    "pinia": "^3.0.2",
     "tailwindcss": "^4.1.5",
     "vue": "^3.5.13",
     "vue-router": "^4.5.1",
