+++ pages/error.vue
@@ -1,0 +1,16 @@
+<script lang="ts" setup>
+const route = useRoute();
+const error = route.query.error || "An unknown error occurred.";
+</script>
+
+<template>
+  <div class="card bg-base-300 container mt-4 min-h-72 text-center mx-auto flex flex-col justify-center items-center gap-4">
+    <div role="alert" class="alert alert-error min-w-1/2 text-center">
+      <span>Error! Task failed successfully. {{ error }}</span>
+    </div>
+    <NuxtLink to="/" class="btn btn-primary">
+      Go back to home
+      <Icon name="tabler:arrow-back-up" size="24" />
+    </NuxtLink>
+  </div>
+</template>
